+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # 1.5.1 / 2014-02-23
+* [FEATURE] Add created_at to boards and items.
 * [FEATURE] Add Board#items
 
 # 1.5.0 / 2014-02-22